增加了文章草稿箱功能，删除文章功能，还完善了搜素功能以及profile页面

+++ src/main/java/com/codeplayer/elasticsearch/ArticleRepository.java
@@ -3,5 +3,5 @@ package com.codeplayer.elasticsearch;
 import com.codeplayer.entity.Article;
 import org.springframework.data.elasticsearch.repository.ElasticsearchRepository;
 
-public interface ArticleRepository extends ElasticsearchRepository <Article,Integer>{
+public interface ArticleRepository extends ElasticsearchRepository <Article,Long>{
 }
